wasm glclear working
