+++ examples/rust-webpack/crate/Cargo.toml
@@ -1,11 +1,12 @@
 [package]
-authors = ["The RustWasm Developers"]
+authors = []
 categories = ["wasm"]
 description = "My super awesome Rust, WebAssembly, and Webpack project!"
 license = "Apache-2.0/MIT"
 name = "rust-webpack"
 readme = "./README.md"
 repository = "https://github.com/bwasty/yage"
+edition = "2018"
 version = "0.1.0"
 
 [lib]
