+++ yage-gl/src/lib.rs
@@ -1,9 +1,9 @@
 #[cfg(not(target_arch = "wasm32"))]
-#[path = "gl.rs"]
+#[path = "gl_native.rs"]
 mod gl;
 
 #[cfg(target_arch = "wasm32")]
-#[path = "webgl.rs"]
+#[path = "gl_web.rs"]
 pub mod gl;
 
 pub use crate::gl::*;
